+++ database/migrations/2017_05_02_092838_create_slots_table.php
@@ -12,41 +12,48 @@ class CreateSlotsTable extends Migration
      */
     public function up()
     {
-        Schema::create('slots', function (Blueprint $table) {
-            $table->increments('id');
-
-            $table->integer('time_tp')->default(0);
-            $table->integer('time_td')->default(0);
-            $table->integer('time_course')->default(0);
-            $table->integer('course_annual_id')->unsigned();
-            $table->integer('course_session_id')->unsigned();
-            $table->integer('lecturer_id')->unsigned()->nullable();
-            $table->integer('responsible_department_id')->unsigned()->nullable();
-            $table->integer('group_id')->unsigned()->nullable();
-            $table->double('time_used')->nullable();
-            $table->double('time_remaining')->nullable();
-
-            $table->integer('created_uid')->unsigned();
-            $table->integer('write_uid')->unsigned()->nullable();
 
 
-            $table->foreign('course_annual_id')
-                ->references('id')
-                ->on('course_annuals')
-                ->onDelete('cascade');
+        if(!Schema::hasTable('slots')) {
 
-            $table->foreign('course_session_id')
-                ->references('id')
-                ->on('course_sessions')
-                ->onDelete('cascade');
+            Schema::create('slots', function (Blueprint $table) {
+                $table->increments('id');
 
-            $table->foreign('group_id')
-                ->references('id')
-                ->on('groups')
-                ->onDelete('cascade');
+                $table->integer('time_tp')->default(0);
+                $table->integer('time_td')->default(0);
+                $table->integer('time_course')->default(0);
+                $table->integer('course_annual_id')->unsigned();
+                $table->integer('course_session_id')->unsigned();
+                $table->integer('lecturer_id')->unsigned()->nullable();
+                $table->integer('responsible_department_id')->unsigned()->nullable();
+                $table->integer('group_id')->unsigned()->nullable();
+                $table->double('time_used')->nullable();
+                $table->double('time_remaining')->nullable();
 
-            $table->timestamps();
-        });
+                $table->integer('created_uid')->unsigned();
+                $table->integer('write_uid')->unsigned()->nullable();
+
+
+                $table->foreign('course_annual_id')
+                    ->references('id')
+                    ->on('course_annuals')
+                    ->onDelete('cascade');
+
+                $table->foreign('course_session_id')
+                    ->references('id')
+                    ->on('course_sessions')
+                    ->onDelete('cascade');
+
+                $table->foreign('group_id')
+                    ->references('id')
+                    ->on('groups')
+                    ->onDelete('cascade');
+
+                $table->timestamps();
+            });
+
+        }
+
 
         // Remove time_used and time_remaining columns from course_sessions table.
        /* Schema::table('course_sessions', function (Blueprint $table) {
