+++ database/migrations/2017_05_02_092839_create_timetable_slots_table.php
@@ -12,53 +12,59 @@ class CreateTimetableSlotsTable extends Migration
      */
     public function up()
     {
-        Schema::create('timetable_slots', function (Blueprint $table) {
-            $table->increments('id');
 
-            $table->integer('timetable_id')->unsigned();
-            $table->integer('course_session_id')->unsigned();
-            $table->integer('slot_id')->unsigned();
-            $table->integer('room_id')->nullable();
-            $table->integer('group_merge_id')->unsiged();
+        if(!Schema::hasTable('timetable_slots')) {
 
-            $table->string('course_name');
-            $table->string('teacher_name');
-            $table->string('type');
-            $table->double('durations')->unsigned();
+            Schema::create('timetable_slots', function (Blueprint $table) {
+                $table->increments('id');
 
-            $table->dateTime('start');
-            $table->dateTime('end');
-            $table->integer('created_uid');
-            $table->integer('updated_uid');
+                $table->integer('timetable_id')->unsigned();
+                $table->integer('course_session_id')->unsigned();
+                $table->integer('slot_id')->unsigned();
+                $table->integer('room_id')->nullable();
+                $table->integer('group_merge_id')->unsiged();
 
-            $table->foreign('timetable_id')
-                ->references('id')
-                ->on('timetables')
-                ->onDelete('cascade');
+                $table->string('course_name');
+                $table->string('teacher_name');
+                $table->string('type');
+                $table->double('durations')->unsigned();
 
-            $table->foreign('course_session_id')
-                ->references('id')
-                ->on('course_sessions')
-                ->onDelete('cascade');
+                $table->dateTime('start');
+                $table->dateTime('end');
+                $table->integer('created_uid');
+                $table->integer('updated_uid');
 
-            $table->foreign('room_id')
-                ->references('id')
-                ->on('rooms')
-                ->onDelete('cascade');
+                $table->foreign('timetable_id')
+                    ->references('id')
+                    ->on('timetables')
+                    ->onDelete('cascade');
 
-            $table->foreign('group_merge_id')
-                ->references('id')
-                ->on('merge_timetable_slots')
-                ->onDelete('cascade');
+                $table->foreign('course_session_id')
+                    ->references('id')
+                    ->on('course_sessions')
+                    ->onDelete('cascade');
 
-            $table->foreign('slot_id')
-                ->references('id')
-                ->on('slots')
-                ->onDelete('cascade');
+                $table->foreign('room_id')
+                    ->references('id')
+                    ->on('rooms')
+                    ->onDelete('cascade');
 
-            
-            $table->timestamps();
-        });
+                $table->foreign('group_merge_id')
+                    ->references('id')
+                    ->on('merge_timetable_slots')
+                    ->onDelete('cascade');
+
+                $table->foreign('slot_id')
+                    ->references('id')
+                    ->on('slots')
+                    ->onDelete('cascade');
+
+
+                $table->timestamps();
+            });
+
+        }
+
     }
 
     /**
